+++ packages/xchain-client/CHANGELOG.md
@@ -5,7 +5,7 @@
 - Use latest `@xchainjs/xchain-util@0.3.2`
 - made walletIndex in BaseXChainClient.getAddress(walletIndex?: number)
 - upgrade axios dependency to latest
-- change TX.asset to allow Asset[]
+- change TxFrom/TxTo to have optional Asset, to support Terra's multiple native asset types (UST, KRT, LUNA, etc)
 
 # v.0.11.0 (2021-12-29)